+++ shared/js/file/LocalIcons.ts
@@ -150,6 +150,8 @@ class IconManager extends AbstractIconManager {
     }
 
     resolveIcon(iconId: number, serverUniqueId: string, handlerIdHint: string): RemoteIcon {
+        serverUniqueId = serverUniqueId || "global";
+
         /* just to ensure */
         iconId = iconId >>> 0;