+++ shared/js/ui/react-elements/ModalDefinitions.ts
@@ -44,6 +44,7 @@ export abstract class AbstractModal {
 
     /* only valid for the "inline" modals */
     type() : ModalType { return "none"; }
+    color() : "none" | "blue" { return "none"; }
 
     protected onInitialize() {}
     protected onDestroy() {}